+++ petanque/json/interp.ml
@@ -65,4 +65,8 @@ let handle_request ~(do_handle : 'a handle) ~unhandled ~token ~method_ ~params =
     do_handle ~token (do_request (module StateEqual) ~params)
   | s when String.equal StateHash.method_ s ->
     do_handle ~token (do_request (module StateHash) ~params)
+  | s when String.equal StateProofEqual.method_ s ->
+    do_handle ~token (do_request (module StateProofEqual) ~params)
+  | s when String.equal StateProofHash.method_ s ->
+    do_handle ~token (do_request (module StateProofHash) ~params)
   | _ -> unhandled ~token ~method_